divide login stack file
